Combine director & opsman config generation
TL;DR
=====
To keep this DRY, the tasks that generate the director & operations
manager were put into the same task.

The platform-automation tasks are expecting a single `vars` input, so
we need to make sure that our configurations exist in the same
directory. As such, we have a single output, but configurations for
operations manager and the BOSH director are both copied to that
directory.