Functional pipeline - typos, proper var assignment
TL;DR
=====
- Fix typos
- Add proper task requisites
- Rename confusing variables
- Fix subnets

This commit produces a functional pipeline for installing Pivotal
Operations Manager and the associated BOSH director. Previously, there
were some unwanted typos and task dependencies that were
non-functional. For example, the old name of `install-opsman` is no
longer relevant, and the `install-opsman-and-director` job has since
replaced it. This is the result of combining the jobs into one, to
reduce the amount of reused code.

In addition to the aforementioned changes, there were some IP
addressing issues that were corrected.