Organize product files, YAML multiline string fix
TL;DR
=====
- Inject `\n` and trim `\r` during YAML generation
- Move products into version directories
- Remove PAS pipeline (not used)

Detail
======
YAML generation (from JSON) would not parse multiline strings correctly,
which only really impacted certificates. This would require certificates
to be placed in Credhub with `\n\r` appended to each newline. This is
now unnecessary, as the variable is being manipulated now with `awk` and
`tr` to add `\n` and truncate `\r` (respectively).

Remove the PAS pipeline. This pipeline was not used, and is counter to
the intent of these pipelines.

Organize products. Products should live in their version-based
directories, rather than updating a single parameter file. This is
opinionated, but also an easy path forward for upgrades.